requirements.txt
Flask==2.2.2
pandas==1.5.0
python-dotenv==0.21.0
requests==2.28.1
